# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Registrar novo usuário",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso"
          },
          "409": {
            "description": "E-mail já cadastrado"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Fazer login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "summary": "Renovar token de acesso",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token renovado com sucesso"
          },
          "401": {
            "description": "Token inválido"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Fazer logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Obter perfil do usuário logado",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "summary": "Listar produtos com filtros e paginação",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brandId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "OUT_OF_STOCK",
                "DISCONTINUED"
              ],
              "type": "string"
            }
          },
          {
            "name": "isFeatured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isOrganic",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos retornada com sucesso"
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "summary": "Criar novo produto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto criado com sucesso"
          },
          "409": {
            "description": "SKU ou código de barras já cadastrado"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "summary": "Obter carrinho do usuário",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Carrinho retornado com sucesso"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CartController_clearCart",
        "summary": "Limpar carrinho",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Carrinho limpo com sucesso"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/cart/items": {
      "post": {
        "operationId": "CartController_addItem",
        "summary": "Adicionar item ao carrinho",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item adicionado com sucesso"
          },
          "400": {
            "description": "Estoque insuficiente"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/cart/items/{itemId}": {
      "patch": {
        "operationId": "CartController_updateItem",
        "summary": "Atualizar quantidade do item no carrinho",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item atualizado com sucesso"
          },
          "400": {
            "description": "Estoque insuficiente"
          },
          "404": {
            "description": "Item não encontrado"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "summary": "Remover item do carrinho",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removido com sucesso"
          },
          "404": {
            "description": "Item não encontrado"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/cart/merge": {
      "post": {
        "operationId": "CartController_mergeGuestCart",
        "summary": "Mesclar carrinho de visitante com carrinho do usuário",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Carrinhos mesclados com sucesso"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/analytics/dashboard": {
      "get": {
        "operationId": "AnalyticsController_getAdminDashboard",
        "summary": "Obter dashboard geral (admin/gerente)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dashboard retornado com sucesso"
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/sales": {
      "get": {
        "operationId": "AnalyticsController_getSalesDashboard",
        "summary": "Obter dados de vendas por período",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados de vendas retornados"
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/products": {
      "get": {
        "operationId": "AnalyticsController_getTopProducts",
        "summary": "Obter produtos mais vendidos",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produtos mais vendidos retornados"
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/customers": {
      "get": {
        "operationId": "AnalyticsController_getCustomerAnalytics",
        "summary": "Obter análise de clientes",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Análise de clientes retornada"
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/revenue": {
      "get": {
        "operationId": "AnalyticsController_getRevenueByDay",
        "summary": "Obter receita diária (dados para gráfico)",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados de receita retornados"
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/analytics/sellers": {
      "get": {
        "operationId": "AnalyticsController_getSellerPerformance",
        "summary": "Obter desempenho dos vendedores",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Desempenho dos vendedores retornado"
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Supermercado API",
    "description": "API completa para plataforma de supermercado - Venâncio Aires, RS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "password": {
            "type": "string",
            "example": "Senha@123"
          },
          "firstName": {
            "type": "string",
            "example": "João"
          },
          "lastName": {
            "type": "string",
            "example": "Silva"
          },
          "phone": {
            "type": "string",
            "example": "51999998888"
          },
          "cpf": {
            "type": "string",
            "example": "12345678901"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "password": {
            "type": "string",
            "example": "Senha@123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "example": "SKU-001"
          },
          "barcode": {
            "type": "string",
            "example": "7891234567890"
          },
          "name": {
            "type": "string",
            "example": "Arroz Integral Orgânico 1kg"
          },
          "description": {
            "type": "string",
            "example": "Arroz integral orgânico de alta qualidade"
          },
          "shortDescription": {
            "type": "string",
            "example": "Arroz integral orgânico"
          },
          "categoryId": {
            "type": "string",
            "example": "uuid-da-categoria"
          },
          "brandId": {
            "type": "string",
            "example": "uuid-da-marca"
          },
          "price": {
            "type": "number",
            "example": 12.9
          },
          "costPrice": {
            "type": "number",
            "example": 8.5
          },
          "compareAtPrice": {
            "type": "number",
            "example": 15.9
          },
          "stock": {
            "type": "number",
            "example": 100,
            "default": 0
          },
          "minStock": {
            "type": "number",
            "example": 5,
            "default": 5
          },
          "maxStock": {
            "type": "number",
            "example": 500
          },
          "unit": {
            "type": "string",
            "example": "kg",
            "default": "un"
          },
          "weight": {
            "type": "number",
            "example": 1
          },
          "volume": {
            "type": "number",
            "example": 0.5
          },
          "mainImageUrl": {
            "type": "string",
            "example": "https://example.com/image.jpg"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "OUT_OF_STOCK",
              "DISCONTINUED"
            ],
            "default": "ACTIVE"
          },
          "isFeatured": {
            "type": "boolean",
            "example": false,
            "default": false
          },
          "isOrganic": {
            "type": "boolean",
            "example": false,
            "default": false
          },
          "taxRate": {
            "type": "number",
            "example": 0,
            "default": 0
          },
          "tags": {
            "example": [
              "orgânico",
              "integral"
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aisleLocation": {
            "type": "string",
            "example": "Corredor 3"
          },
          "shelfPosition": {
            "type": "string",
            "example": "Prateleira B2"
          }
        },
        "required": [
          "sku",
          "name",
          "categoryId",
          "price"
        ]
      },
...
      "AddToCartDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "uuid-do-produto",
            "description": "ID do produto"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Quantidade do produto",
            "minimum": 1
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Nova quantidade do item",
            "minimum": 1
          }
        },
        "required": [
          "quantity"
        ]
      },
...
    }
  }
}